pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cinelog-backend"
version = "0.1.0"
description = "Backend API for CineLog iOS app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "pytest>=8.0",
  "httpx>=0.27",
]

[tool.setuptools]
# Tell setuptools to package the "app" directory as a Python package
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"
